+++ Angular-Front-End/app-front-end/src/app/appSettings.ts
@@ -1,3 +1,4 @@
 export class AppSettings {
-    public static APP_NAME='FCUL App';
+    public static APP_NAME ='Instigo';
+    public static APP_ICON_DIRECTORY ="../../../assets/icons/Instigo.png";
 }