+++ Angular-Front-End/app-front-end/src/app/services/controllers/communities-controller.service.ts
@@ -3,6 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import { TemplateControllerService } from './template-controller.service';
 import { CommunityModel } from 'src/app/models/community.model';
 import { AppRoutesService } from '../router/app-routes.service';
+import { Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -25,23 +26,23 @@ export class CommunitiesService extends TemplateControllerService<CommunityModel
     return this.appRoutes.apiCommunityEndPoint;
   }
 
-  public getCommunityLists(uID: string) {
+  public getCommunityLists(uID: string): Observable<any> {
     return this.http.get(`${this.appRoutes.apiCommunityMainInfo}/${uID}`);
   }
 
-  public subscribeToCommunity(uID: string, cID: string): any {
+  public subscribeToCommunity(uID: string, cID: string): Observable<any> {
     return this.http.post(this.appRoutes.apiUserCommunitySubscribeEndPoint, {uID, cID});
   }
 
-  public unsubscribeFromCommunity(uID: string, cID: string): any {
+  public unsubscribeFromCommunity(uID: string, cID: string): Observable<any> {
     return this.http.post(this.appRoutes.apiUserCommunityUnsubscribeEndPoint, {uID, cID});
   }
 
-  public userSubbedCommunities(uID: string): any {
+  public userSubbedCommunities(uID: string): Observable<any> {
     return this.http.get(`${this.appRoutes.apiUserSubscribedCommunitiesEndPoint}/${uID}`);
   }
 
-  public userCreatedCommunities(ownerUserName: string): any {
+  public userCreatedCommunities(ownerUserName: string): Observable<any> {
     return this.http.get(`${this.appRoutes.apiUserCreatedCommunitiesEndPoint}/${ownerUserName}`);
   }