+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/UserEventCatalog.java
@@ -1,9 +1,11 @@
 package io.App.EventService.EventComponent;
 
+import java.util.List;
+
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 import io.App.EventService.databaseConnection.UserEventDatabaseManagement;
-import io.App.EventService.dto.EventListWrapper;
+import io.App.EventService.dto.EventDTO;
 import io.App.EventService.exceptions.InternalAppException;
 
 @SpringBootApplication
@@ -15,8 +17,12 @@ public class UserEventCatalog {
 		this.uEDM = new UserEventDatabaseManagement();
 	}
 
-	public EventListWrapper eventsFromSubbedCommunities(int uID) throws NumberFormatException, InternalAppException {
+	public List<EventDTO> eventsFromSubbedCommunities(int uID) throws NumberFormatException, InternalAppException {
 		return uEDM.eventsFromSubCommunities(uID);
+	}
+
+	public List<EventDTO> eventsCreatedByUser(String userName) throws InternalAppException {
+		return uEDM.userCreatedEvents(userName);
 	}
 
 }