+++ Services/EventService/src/main/java/io/App/EventService/controller/UserEventController.java
@@ -32,7 +32,7 @@ public class UserEventController {
 		EventListWrapper eLW = null;
 
 		try {
-			eLW = uEC.eventsFromSubbedCommunities(Integer.parseInt(uID));
+			eLW = new EventListWrapper(uEC.eventsFromSubbedCommunities(Integer.parseInt(uID)));
 		} catch (NumberFormatException e) {
 			System.err.println(e.getMessage());
 			return new ResponseEntity<>(new Pair<>(INTERNAL_APP_ERROR_MESSAGE, null),
@@ -43,8 +43,25 @@ public class UserEventController {
 					HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 		
-		System.out.println("Successfull events request");
+		System.out.println("Successfull subscribed events request");
 		return new ResponseEntity<>(new Pair<>("Successfull events request", eLW),
+				HttpStatus.OK);
+	}
+	
+	@GetMapping(path = "/userCreatedEvents/{userName}")
+	public ResponseEntity<Pair<String, EventListWrapper>> userCreatedEvents(@PathVariable("userName") String userName) {
+		EventListWrapper eLW = null;
+		System.out.println(userName);
+		try {
+			eLW = new EventListWrapper(uEC.eventsCreatedByUser(userName));
+		} catch (InternalAppException e) {
+			System.err.println(e.getMessage());
+			return new ResponseEntity<>(new Pair<>(e.getMessage(), null),
+					HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		System.out.println("Successfull created events request");
+		return new ResponseEntity<>(new Pair<>("Successfull created events request", eLW),
 				HttpStatus.OK);
 	}
 }