+++ Services/EventService/src/main/java/io/App/EventService/databaseConnection/UserEventDatabaseManagement.java
@@ -7,9 +7,9 @@ import java.sql.SQLException;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.List;
 
 import io.App.EventService.dto.EventDTO;
-import io.App.EventService.dto.EventListWrapper;
 import io.App.EventService.exceptions.InternalAppException;
 
 public class UserEventDatabaseManagement {
@@ -19,6 +19,7 @@ public class UserEventDatabaseManagement {
 
 	private static final String EVENTS_FROM_SUBSCRIBED_COMMUNITIES_SQL = "SELECT * FROM Events AS E WHERE (E.cID = "
 			+ "(SELECT cID FROM RolesUsersCommunities AS RUC WHERE (RUC.cID = E.cID) AND (RUC.uID = ?)));";
+	private static final String USER_CREATED_EVENTS_SQL = "SELECT * FROM Events WHERE (ownerUserName = ?)";
 
 	public UserEventDatabaseManagement() {
 		databaseConnection = new DatabaseConnection();
@@ -31,12 +32,11 @@ public class UserEventDatabaseManagement {
 	 * @return all events from the user (uID) subscribed communities
 	 * @throws InternalAppException
 	 */
-	public EventListWrapper eventsFromSubCommunities(int uID)
+	public List<EventDTO> eventsFromSubCommunities(int uID)
 			throws InternalAppException {
 		Connection con = databaseConnection.connectToDatabase();
 		PreparedStatement stmt = null;
 		ResultSet rs = null;
-		EventListWrapper eLW = new EventListWrapper();
 		ArrayList<EventDTO> listEvents = new ArrayList<>();
 
 		try {
@@ -90,8 +90,66 @@ public class UserEventDatabaseManagement {
 			}
 		}
 
-		eLW.setList(listEvents);
+		return listEvents;
+	}
 
-		return eLW;
+	public List<EventDTO> userCreatedEvents(String userName) throws InternalAppException {
+		Connection con = databaseConnection.connectToDatabase();
+		PreparedStatement stmt = null;
+		ResultSet rs = null;
+		ArrayList<EventDTO> listEvents = new ArrayList<>();
+
+		try {
+			stmt = con.prepareStatement(USER_CREATED_EVENTS_SQL);
+			stmt.setString(1, userName);
+			rs = stmt.executeQuery();
+			while (rs.next()) {
+				String eID = String.valueOf(rs.getInt(1));
+				String eName = rs.getString(2);
+
+				// format dates
+				DateFormat df = new SimpleDateFormat("dd-MM-yyyy");
+				String dateStart = df.format(rs.getDate(3));
+				String dateEnd = df.format(rs.getDate(4));
+
+				String cID = String.valueOf(rs.getInt(5));
+				String cName = rs.getString(6);
+				String ownerUserName = rs.getString(7);
+
+				// create Event
+				EventDTO e = new EventDTO(eID, eName, dateStart, dateEnd, cID,
+						cName, ownerUserName);
+
+				// insert Event into List
+				listEvents.add(e);
+			}
+		} catch (SQLException e) {
+			System.err.println(e.getMessage());
+			throw new InternalAppException();
+		} finally {
+			if (con != null) {
+				try {
+					con.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (stmt != null) {
+				try {
+					stmt.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+			if (rs != null) {
+				try {
+					rs.close();
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			}
+		}
+
+		return listEvents;
 	}
 }